% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{create_mating_matrices}
\alias{create_mating_matrices}
\title{Create Mating Matrices for Breeding Schemes}
\usage{
create_mating_matrices()
}
\value{
A list containing three 3x3 transition matrices:
\item{backcross_AA}{Matrix for backcrossing to AA donor}
\item{backcross_aa}{Matrix for backcrossing to aa donor}
\item{selfing}{Matrix for self-fertilization}
}
\description{
Creates transition matrices for backcrossing and selfing operations
used in Near Isogenic Line (NIL) development.
}
\details{
Each matrix represents genotype transitions where:
\itemize{
\item Rows represent offspring genotypes (AA, Aa, aa)
\item Columns represent parent genotypes (AA, Aa, aa)
\item Values are transition probabilities
}
}
\examples{
matrices <- create_mating_matrices()
matrices$selfing  # Shows selfing transition probabilities

}
