% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{nil_frequencies_for_hmm}
\alias{nil_frequencies_for_hmm}
\title{Convert NIL Frequencies to HMM Format}
\usage{
nil_frequencies_for_hmm(
  bc = 2,
  s = 3,
  donor_type = "aa",
  pop = c(0, 1, 0),
  ...
)
}
\arguments{
\item{bc}{Integer. Number of backcross generations (default: 2)}

\item{s}{Integer. Number of selfing generations (default: 3)}

\item{donor_type}{Character. Type of donor parent, either "AA" or "aa" (default: "aa")}

\item{pop}{Numeric vector of length 3. Initial population frequencies (default: c(0,1,0))}

\item{...}{Additional arguments passed to \code{calculate_nil_frequencies()}}
}
\value{
Named numeric vector with frequencies for REF, HET, ALT states
}
\description{
Converts genotype frequencies from breeding scheme calculations to the
REF/HET/ALT format used in HMM ancestry analysis.
}
\details{
This function maps biological genotypes to ancestry states:
\itemize{
\item REF: Recurrent parent background
\item HET: Heterozygous regions (admixed)
\item ALT: Donor introgression regions
}

The mapping depends on donor_type:
\itemize{
\item If donor_type = "aa": REF=AA, HET=Aa, ALT=aa
\item If donor_type = "AA": REF=aa, HET=Aa, ALT=AA
}
}
\examples{
# BC2S3 with teosinte (aa) donor
bc2s3_hmm <- nil_frequencies_for_hmm(bc = 2, s = 3, donor_type = "aa")
print(bc2s3_hmm)

}
